Simplify solution and unit-tests
