+++ __tests__/util.test.ts
@@ -1,6 +1,10 @@
 import * as cache from '@actions/cache';
 import * as core from '@actions/core';
-import {isVersionSatisfies, isCacheFeatureAvailable} from '../src/util';
+import {
+  convertVersionToSemver,
+  isVersionSatisfies,
+  isCacheFeatureAvailable
+} from '../src/util';
 
 jest.mock('@actions/cache');
 jest.mock('@actions/core');
@@ -61,5 +65,18 @@ describe('isCacheFeatureAvailable', () => {
   it('isCacheFeatureAvailable is enabled', () => {
     jest.spyOn(cache, 'isFeatureAvailable').mockImplementation(() => true);
     expect(isCacheFeatureAvailable()).toBe(true);
+  });
+});
+
+describe('convertVersionToSemver', () => {
+  it.each([
+    ['12', '12'],
+    ['12.0', '12.0'],
+    ['12.0.2', '12.0.2'],
+    ['12.0.2.1', '12.0.2+1'],
+    ['12.0.2.1.0', '12.0.2+1.0']
+  ])('%s -> %s', (input: string, expected: string) => {
+    const actual = convertVersionToSemver(input);
+    expect(actual).toBe(expected);
   });
 });
